requirements.txt
python-socketio~==4.6.0
